requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==24.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
python-decouple==3.8
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0
qrcode==7.4.2
reportlab==4.0.9
openpyxl==3.1.2